Fixed warnings.

+++ examples/12-lod/lod.cpp
@@ -146,7 +146,7 @@ class ExampleLod : public entry::AppI
 			imguiEndFrame();
 
 			// Set view 0 default viewport.
-			bgfx::setViewRect(0, 0, 0, m_width, m_height);
+			bgfx::setViewRect(0, 0, 0, uint16_t(m_width), uint16_t(m_height) );
 
 			// This dummy draw call is here to make sure that view 0 is cleared
 			// if no other draw calls are submitted to view 0.
